refactor(demo): split receive page widgets
